Example Data: Sparrow Repeatability (3 Point Data)

An example of data that one would perform profile repeatability on. Consists of 12 individual animals, with data taken at 3 times (n_trials = 3). Then, there are 10 replicate columns. This example also shows what happens when there are null data records for some individuals.

Usage

sparrow_repeatability_three_point

Format

sparrow_repeatability_three_point

A dataframe with 36 rows and 12 columns:

Animal

The animal name/unique identifier

TIME

The time of the measurement, in days

LD.500

The name of the replicate column

Source

An experiment run at Tufts University by Ursula K. Beattie